feat: rename redis client file - add missing types
