+++ jest.setup.js
@@ -1,2 +1,4 @@
 // Learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
+
+process.env.REDIS_URL = 'redis://localhost:6379';
