+++ package.json
@@ -5,7 +5,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test": "jest --watch"
+    "test": "jest",
+    "test:watch": "jest --watch"
   },
   "dependencies": {
     "@apollo/client": "^3.7.1",
@@ -31,6 +32,7 @@
     "@types/node": "^18.11.9",
     "@types/react": "^18.2.34",
     "@types/react-dom": "^18.2.14",
+    "@types/uuid": "^9.0.7",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "ts-jest": "^29.1.1",
