--- src/libs/index.ts
@@ -1,2 +1,0 @@
-export * from './redis';
-export * from './mdx';
